Merge remote-tracking branch 'ardupilot/master'

Conflicts:
	ArduCopter/APM_Config.h
	ArduCopter/ArduCopter.pde
	ArduCopter/GCS_Mavlink.pde
	ArduCopter/Parameters.pde
	ArduCopter/compat.pde
	ArduCopter/config.h
	ArduCopter/control_poshold.pde
	ArduCopter/perf_info.pde
	ArduCopter/system.pde
	libraries/AP_AHRS/AP_AHRS_DCM.h
	libraries/AP_Compass/AP_Compass.h
	libraries/AP_Compass/Compass.h
	libraries/AP_HAL/AP_HAL_Boards.h
	libraries/AP_NavEKF/AP_NavEKF.cpp

+++ libraries/AC_PID/AC_PI_2D.h
@@ -50,7 +50,7 @@ public:
     void        save_gains();
 
     /// operator function call for easy initialisation
-    void operator() (float p, float i, float imaxval, float input_filt_hz, float dt );
+    void operator() (float p, float i, float imaxval, float input_filt_hz, float dt);
 
     // get accessors
     float       kP() const { return _kp.get(); }
@@ -62,7 +62,7 @@ public:
     // set accessors
     void        kP(const float v) { _kp.set(v); }
     void        kI(const float v) { _ki.set(v); }
-    void        imax(const float v) { _imax.set(fabs(v)); }
+    void        imax(const float v) { _imax.set(fabsf(v)); }
     void        filt_hz(const float v);
 
     Vector2f    get_integrator() const { return _integrator; }
